# artifacts/api-server/src/routes/gst.ts
import { Router } from "express";
import { db, invoicesTable, vendorBillsTable } from "@workspace/db";
import { gte, lte, and, eq } from "drizzle-orm";

const router = Router();

function parsePeriod(period: string) {
  const [month, year] = period.split("-");
  const months: Record<string, string> = {
    Jan: "01", Feb: "02", Mar: "03", Apr: "04", May: "05", Jun: "06",
    Jul: "07", Aug: "08", Sep: "09", Oct: "10", Nov: "11", Dec: "12",
  };
  const m = months[month] || "01";
  const y = year || new Date().getFullYear().toString();
  const from = `${y}-${m}-01`;
  const lastDay = new Date(Number(y), Number(m), 0).getDate();
  const to = `${y}-${m}-${String(lastDay).padStart(2, "0")}`;
  return { from, to };
}

router.get("/gst/itc-ledger", async (req, res) => {
  try {
    const period = (req.query.period as string) || "May-2025";
    const { from, to } = parsePeriod(period);
    const bills = await db.select().from(vendorBillsTable)
      .where(and(gte(vendorBillsTable.date, from), lte(vendorBillsTable.date, to)));

    const cgstAvailed = bills.reduce((s, b) => s + Number(b.cgst), 0);
    const sgstAvailed = bills.reduce((s, b) => s + Number(b.sgst), 0);
    const igstAvailed = bills.reduce((s, b) => s + Number(b.igst), 0);

    res.json({
      period,
      openingBalance: { cgst: 18000, sgst: 18000, igst: 8000, total: 44000 },
      availed: { cgst: cgstAvailed, sgst: sgstAvailed, igst: igstAvailed, total: cgstAvailed + sgstAvailed + igstAvailed },
      reversed: { cgst: 0, sgst: 0, igst: 0, total: 0 },
      utilized: { cgst: 12000, sgst: 12000, igst: 5000, total: 29000 },
      closingBalance: {
        cgst: 18000 + cgstAvailed - 12000,
        sgst: 18000 + sgstAvailed - 12000,
        igst: 8000 + igstAvailed - 5000,
        total: 44000 + cgstAvailed + sgstAvailed + igstAvailed - 29000,
      },
    });
  } catch (err) {
    req.log.error(err);
    res.status(500).json({ error: "Failed to get ITC ledger" });
  }
});

router.get("/gst/gstr1", async (req, res) => {
  try {
    const period = (req.query.period as string) || "May-2025";
    const { from, to } = parsePeriod(period);
    const invoices = await db.select().from(invoicesTable)
      .where(and(gte(invoicesTable.date, from), lte(invoicesTable.date, to), eq(invoicesTable.status, "posted")));

    const b2b = invoices.filter(i => i.customerGstin);
    const b2c = invoices.filter(i => !i.customerGstin);

    const totalTaxableValue = invoices.reduce((s, i) => s + Number(i.taxableAmount), 0);
    const totalCgst = invoices.reduce((s, i) => s + Number(i.cgst), 0);
    const totalSgst = invoices.reduce((s, i) => s + Number(i.sgst), 0);
    const totalIgst = invoices.reduce((s, i) => s + Number(i.igst), 0);

    res.json({
      period,
      totalInvoices: invoices.length,
      b2bAmount: b2b.reduce((s, i) => s + Number(i.totalAmount), 0),
      b2cAmount: b2c.reduce((s, i) => s + Number(i.totalAmount), 0),
      exportAmount: 0,
      creditNoteAmount: invoices.filter(i => i.type === "Credit Note").reduce((s, i) => s + Number(i.totalAmount), 0),
      totalTaxableValue,
      totalCgst,
      totalSgst,
      totalIgst,
      totalTax: totalCgst + totalSgst + totalIgst,
      isFilingReady: invoices.length > 0,
      validationErrors: [],
    });
  } catch (err) {
    req.log.error(err);
    res.status(500).json({ error: "Failed to get GSTR-1" });
  }
});

router.get("/gst/gstr3b", async (req, res) => {
  try {
    const period = (req.query.period as string) || "May-2025";
    const { from, to } = parsePeriod(period);

    const invoices = await db.select().from(invoicesTable)
      .where(and(gte(invoicesTable.date, from), lte(invoicesTable.date, to), eq(invoicesTable.status, "posted")));
    const bills = await db.select().from(vendorBillsTable)
      .where(and(gte(vendorBillsTable.date, from), lte(vendorBillsTable.date, to)));

    const outwardTaxable = invoices.reduce((s, i) => s + Number(i.taxableAmount), 0);
    const outwardCgst = invoices.reduce((s, i) => s + Number(i.cgst), 0);
    const outwardSgst = invoices.reduce((s, i) => s + Number(i.sgst), 0);
    const outwardIgst = invoices.reduce((s, i) => s + Number(i.igst), 0);
    const outwardTax = outwardCgst + outwardSgst + outwardIgst;

    const itcCgst = bills.reduce((s, b) => s + Number(b.cgst), 0);
    const itcSgst = bills.reduce((s, b) => s + Number(b.sgst), 0);
    const itcIgst = bills.reduce((s, b) => s + Number(b.igst), 0);
    const totalItc = itcCgst + itcSgst + itcIgst;

    const netPayable = Math.max(0, outwardTax - totalItc);

    res.json({
      period,
      outwardSupplies: outwardTaxable,
      outwardTax,
      inwardRcm: 0,
      itcAvailable: { cgst: itcCgst, sgst: itcSgst, igst: itcIgst, total: totalItc },
      itcReversed: { cgst: 0, sgst: 0, igst: 0, total: 0 },
      netItc: { cgst: itcCgst, sgst: itcSgst, igst: itcIgst, total: totalItc },
      taxPayableCash: netPayable,
      taxPayableItc: totalItc,
      totalLiability: outwardTax,
    });
  } catch (err) {
    req.log.error(err);
    res.status(500).json({ error: "Failed to get GSTR-3B" });
  }
});

router.get("/gst/reconciliation", async (req, res) => {
  try {
    const period = (req.query.period as string) || "May-2025";
    const { from, to } = parsePeriod(period);
    const bills = await db.select().from(vendorBillsTable)
      .where(and(gte(vendorBillsTable.date, from), lte(vendorBillsTable.date, to)));

    const lines = bills.map(b => ({
      vendorGstin: "29AABCT1332L1ZT",
      vendorName: b.vendorName,
      invoiceNo: b.vendorInvoiceNo || b.billNo,
      invoiceDate: b.date,
      bookAmount: Number(b.cgst) + Number(b.sgst) + Number(b.igst),
      gstr2bAmount: null,
      difference: Number(b.cgst) + Number(b.sgst) + Number(b.igst),
      status: "unmatched",
    }));

    res.json({
      period,
      totalInPurchaseRegister: bills.length,
      totalIn2b: Math.max(0, bills.length - 2),
      matched: Math.max(0, bills.length - 2),
      unmatched: 2,
      mismatched: 0,
      unmatchedItcAmount: lines.filter(l => l.status === "unmatched").reduce((s, l) => s + l.bookAmount, 0),
      lines,
    });
  } catch (err) {
    req.log.error(err);
    res.status(500).json({ error: "Failed to get GST reconciliation" });
  }
});

export default router;