# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mystics Audit — Cloud-Based Accounting & GST Compliance SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expense claims
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: employeeId, in: query, schema: { type: integer } }
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
      responses:
        "200":
          description: List of expense claims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseClaim"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Submit expense claim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseClaim"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get expense claim
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Expense claim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseClaim"
  /expenses/{id}/approve:
    post:
      operationId: approveExpense
      tags: [expenses]
      summary: Approve or reject expense claim
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "200":
          description: Approval recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseClaim"
  /expenses/analytics:
    get:
      operationId: getExpenseAnalytics
      tags: [expenses]
      summary: Expense analytics by department and category
      parameters:
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
      responses:
        "200":
          description: Expense analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseAnalytics"

  # PURCHASE ORDERS
...
  /inventory/items:
    get:
      operationId: listItems
      tags: [inventory]
      summary: List inventory items
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string } }
        - { name: group, in: query, schema: { type: string } }
        - { name: lowStock, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createItem
      tags: [inventory]
      summary: Create inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/items/{id}:
    get:
      operationId: getItem
      tags: [inventory]
      summary: Get item with stock position
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    patch:
      operationId: updateItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
  /gst/itc-ledger:
    get:
      operationId: getItcLedger
      tags: [gst]
      summary: Input tax credit ledger (CGST/SGST/IGST)
      parameters:
        - { name: period, in: query, schema: { type: string } }
      responses:
        "200":
          description: ITC ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItcLedger"
  /gst/gstr1:
    get:
      operationId: getGstr1Data
      tags: [gst]
      summary: GSTR-1 return data for a period
      parameters:
        - { name: period, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: GSTR-1 data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gstr1Data"
  /gst/gstr3b:
    get:
      operationId: getGstr3bData
      tags: [gst]
      summary: GSTR-3B return data for a period
      parameters:
        - { name: period, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: GSTR-3B data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gstr3bData"
  /gst/reconciliation:
    get:
      operationId: getGstReconciliation
      tags: [gst]
      summary: GSTR-2B vs purchase register reconciliation
      parameters:
        - { name: period, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Reconciliation data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GstReconciliation"

  # BUDGETS
...
components:
  schemas:
...
    ItcBalance:
      type: object
      properties:
        cgst: { type: number }
        sgst: { type: number }
        igst: { type: number }
        total: { type: number }
      required: [cgst, sgst, igst, total]
...
    ExpenseClaim:
      type: object
      properties:
        id: { type: integer }
        claimNo: { type: string }
        employeeId: { type: integer }
        employeeName: { type: string }
        submittedDate: { type: string }
        totalAmount: { type: number }
        approvedAmount: { type: ["number", "null"] }
        status: { type: string }
        currentApprover: { type: ["string", "null"] }
        policyViolations: { type: integer }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        lines: { type: array, items: { $ref: "#/components/schemas/ExpenseLine" } }
      required: [id, claimNo, employeeId, employeeName, submittedDate, totalAmount, status, policyViolations, createdAt, lines]
    ExpenseLine:
      type: object
      properties:
        id: { type: integer }
        date: { type: string }
        category: { type: string }
        subCategory: { type: ["string", "null"] }
        amount: { type: number }
        currency: { type: string }
        vendorName: { type: ["string", "null"] }
        description: { type: string }
        receiptUrl: { type: ["string", "null"] }
        gstAmount: { type: number }
        vendorGstin: { type: ["string", "null"] }
        policyViolation: { type: boolean }
        violationReason: { type: ["string", "null"] }
      required: [id, date, category, amount, currency, description, gstAmount, policyViolation]
    ExpenseInput:
      type: object
      properties:
        notes: { type: string }
        lines: { type: array, items: { $ref: "#/components/schemas/ExpenseLineInput" } }
      required: [lines]
    ExpenseLineInput:
      type: object
      properties:
        date: { type: string }
        category: { type: string }
        subCategory: { type: string }
        amount: { type: number }
        currency: { type: string }
        vendorName: { type: string }
        description: { type: string }
        gstAmount: { type: number }
        vendorGstin: { type: string }
      required: [date, category, amount, currency, description]
    ApprovalInput:
      type: object
      properties:
        action: { type: string }
        comment: { type: string }
        adjustedAmount: { type: number }
      required: [action]
    ExpenseAnalytics:
      type: object
      properties:
        period: { type: string }
        totalAmount: { type: number }
        totalApproved: { type: number }
        totalPending: { type: number }
        byCategory: { type: array, items: { $ref: "#/components/schemas/ExpenseCategoryBreakdown" } }
        byDepartment: { type: array, items: { $ref: "#/components/schemas/ExpenseDeptBreakdown" } }
        policyViolations: { type: integer }
      required: [period, totalAmount, totalApproved, totalPending, byCategory, byDepartment, policyViolations]
    ExpenseCategoryBreakdown:
      type: object
      properties:
        category: { type: string }
        amount: { type: number }
        count: { type: integer }
        pct: { type: number }
      required: [category, amount, count, pct]
    ExpenseDeptBreakdown:
      type: object
      properties:
        department: { type: string }
        amount: { type: number }
        budget: { type: number }
        utilizationPct: { type: number }
      required: [department, amount, budget, utilizationPct]

    # PURCHASE ORDER
...
    InventoryItem:
      type: object
      properties:
        id: { type: integer }
        itemCode: { type: string }
        name: { type: string }
        type: { type: string }
        group: { type: ["string", "null"] }
        hsnSac: { type: string }
        unit: { type: string }
        gstRate: { type: number }
        purchaseRate: { type: ["number", "null"] }
        sellingRate: { type: ["number", "null"] }
        mrp: { type: ["number", "null"] }
        reorderLevel: { type: ["number", "null"] }
        reorderQty: { type: ["number", "null"] }
        valuationMethod: { type: string }
        currentStock: { type: number }
        stockValue: { type: number }
        isActive: { type: boolean }
        isBatchTracked: { type: boolean }
        isSerialTracked: { type: boolean }
        isExpiryTracked: { type: boolean }
        createdAt: { type: string }
      required: [id, itemCode, name, type, hsnSac, unit, gstRate, valuationMethod, currentStock, stockValue, isActive, isBatchTracked, isSerialTracked, isExpiryTracked, createdAt]
    InventoryItemInput:
      type: object
      properties:
        itemCode: { type: string }
        name: { type: string }
        type: { type: string }
        group: { type: string }
        hsnSac: { type: string }
        unit: { type: string }
        gstRate: { type: number }
        purchaseRate: { type: number }
        sellingRate: { type: number }
        mrp: { type: number }
        reorderLevel: { type: number }
        reorderQty: { type: number }
        valuationMethod: { type: string }
        isBatchTracked: { type: boolean }
        isSerialTracked: { type: boolean }
        isExpiryTracked: { type: boolean }
      required: [name, type, hsnSac, unit, gstRate, valuationMethod]
    InventoryItemUpdate:
      type: object
      properties:
        name: { type: string }
        sellingRate: { type: number }
        purchaseRate: { type: number }
        reorderLevel: { type: number }
        reorderQty: { type: number }
        isActive: { type: boolean }
...
    ItcLedger:
      type: object
      properties:
        period: { type: string }
        openingBalance: { $ref: "#/components/schemas/ItcBalance" }
        availed: { $ref: "#/components/schemas/ItcBalance" }
        reversed: { $ref: "#/components/schemas/ItcBalance" }
        utilized: { $ref: "#/components/schemas/ItcBalance" }
        closingBalance: { $ref: "#/components/schemas/ItcBalance" }
      required: [period, openingBalance, availed, reversed, utilized, closingBalance]
    Gstr1Data:
      type: object
      properties:
        period: { type: string }
        totalInvoices: { type: integer }
        b2bAmount: { type: number }
        b2cAmount: { type: number }
        exportAmount: { type: number }
        creditNoteAmount: { type: number }
        totalTaxableValue: { type: number }
        totalCgst: { type: number }
        totalSgst: { type: number }
        totalIgst: { type: number }
        totalTax: { type: number }
        isFilingReady: { type: boolean }
        validationErrors: { type: array, items: { type: string } }
      required: [period, totalInvoices, b2bAmount, b2cAmount, exportAmount, creditNoteAmount, totalTaxableValue, totalCgst, totalSgst, totalIgst, totalTax, isFilingReady, validationErrors]
    Gstr3bData:
      type: object
      properties:
        period: { type: string }
        outwardSupplies: { type: number }
        outwardTax: { type: number }
        inwardRcm: { type: number }
        itcAvailable: { $ref: "#/components/schemas/ItcBalance" }
        itcReversed: { $ref: "#/components/schemas/ItcBalance" }
        netItc: { $ref: "#/components/schemas/ItcBalance" }
        taxPayableCash: { type: number }
        taxPayableItc: { type: number }
        totalLiability: { type: number }
      required: [period, outwardSupplies, outwardTax, inwardRcm, itcAvailable, itcReversed, netItc, taxPayableCash, taxPayableItc, totalLiability]
    GstReconciliation:
      type: object
      properties:
        period: { type: string }
        totalInPurchaseRegister: { type: integer }
        totalIn2b: { type: integer }
        matched: { type: integer }
        unmatched: { type: integer }
        mismatched: { type: integer }
        unmatchedItcAmount: { type: number }
        lines: { type: array, items: { $ref: "#/components/schemas/ReconciliationLine" } }
      required: [period, totalInPurchaseRegister, totalIn2b, matched, unmatched, mismatched, unmatchedItcAmount, lines]
    ReconciliationLine:
      type: object
      properties:
        vendorGstin: { type: string }
        vendorName: { type: string }
        invoiceNo: { type: string }
        invoiceDate: { type: string }
        bookAmount: { type: number }
        gstr2bAmount: { type: ["number", "null"] }
        difference: { type: number }
        status: { type: string }
      required: [vendorGstin, vendorName, invoiceNo, invoiceDate, bookAmount, difference, status]

    # BUDGET
...