# lib/api-zod/src/generated/types/itcBalance.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Mystics Audit — Cloud-Based Accounting & GST Compliance SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface ItcBalance {
  cgst: number;
  sgst: number;
  igst: number;
  total: number;
}